// src/savvy/enums/CommonLabel.ts
/*
 * @Description: CommonLabel
 * @Author: Hao Xiao
 * @Date: 2021-07-19 14:18:01
 * @LastEditTime: 2022-08-16 13:48:54
 * @LastEditors: Aimee Zhang
 */

export enum CommonLabel {
    DASHBOARD = 'Dashboard',
    MY_VISIT = 'My Visits',
    LEAD = 'Lead',
    SCHEDULE = 'Schedule',
    MY_TEAM = 'My Team',
    MY_CUSTOMERS = 'My Customers',
    EXPLORE = 'Explore',
    DELIVERIES = 'Delivery',
    NOTIFICATIONS = 'Notifications',
    SALES_NOTIFICATIONS = 'Sales Notifications',
    SYNC = 'Sync',
    COPILOT = 'Copilot',
    RANDOM_NUM_MIN = 30,
    RANDOM_NUM_MAX = 50,
    MIDNIGHT = '12:00 AM',
    ELEVEN_FORTY_FIVE = '11:45 PM',
    ELEVEN_THIRTY = '11:30 PM',
    FIFTEEN_MINUTE = 15,
    THIRTY_MINUTE = 30,
    SIXTY_MINUTE = 60,
    ERROR_TIP = 'error',
    LOCATION = 'Location',
    ADD_ATTENDEES = 'AddAttendees',
    MEETING = 'Meeting',
    LIGHT = 'light',
    NUMBER_EIGHT = 8,
    NUMBER_SIX = 6,
    NUMBER_MINUS_ONE = -1,
    NUMBER_ONE = 1,
    NUMBER_ZERO = 0,
    NUMBER_TEN = 10,
    NUMBER_ZERO_POINT_TWO = 0.2,
    CUSTOMER_ITEM_CLICK_LOCATION = 'location',
    CUSTOMER_ITEM_CLICK_MODAL = 'item',
    MY_DAY = 'My Day',
    WK_LABEL = 'WK ',
    NOTIFICATION_PREFERENCE = 'Manager-Incomplete Visit;Manager-Gap Time;Manager-Ad-Hoc Visit;Merchandiser-Ad-Hoc Visit;Merchandiser-Shipment;Merchandiser-Remove Visit',
    SWIPE_BTN_WIDTH = 98,
    END_OF_DAY_TIME = ' 23:59:59',
    LEADS = 'Leads'
}